+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * drop support for R 4.2;
 * documentation improvements;
+* fix a failing test when using `testthat > 3.2.3`;
 
 # SpaDES.tools 2.0.7